+++ model/trade_split.go
@@ -2,15 +2,15 @@ package model
 
 // SplitReceive 分账接收方
 type SplitReceive struct {
-	MerchantId int `json:"merchantId"` // 分账商户号
-	Amount     int `json:"amount"`     // 分账金额 单位分
+	MerchantId int64 `json:"merchantId"` // 分账商户号
+	Amount     int64 `json:"amount"`     // 分账金额 单位分
 }
 
 // TradeSplitReq 交易分账
 type TradeSplitReq struct {
 	SN               string         `json:"sn,omitempty"`           // 设备TUSN号,有固定TUSN号可使用sn参数交易
-	MerchantId       int            `json:"merchantId,omitempty"`   // 平台商户号,渠道对接可使用merchantId和terminalId参数交易
-	TerminalId       int            `json:"terminalId,omitempty"`   // 平台终端号,渠道对接可使用merchantId和terminalId参数交易
+	MerchantId       int64          `json:"merchantId,omitempty"`   // 平台商户号,渠道对接可使用merchantId和terminalId参数交易
+	TerminalId       int64          `json:"terminalId,omitempty"`   // 平台终端号,渠道对接可使用merchantId和terminalId参数交易
 	TradeId          int64          `json:"tradeId"`                // 交易流水号
 	SplitReceiveList []SplitReceive `json:"splitReceiveList"`       // 分账接收方列表
 	SplitRemarks     string         `json:"splitRemarks,omitempty"` // 分账备注
@@ -19,8 +19,8 @@ type TradeSplitReq struct {
 // TradeSplitQueryReq 交易分账查询
 type TradeSplitQueryReq struct {
 	SN           string `json:"sn,omitempty"`         // 设备TUSN号,有固定TUSN号可使用sn参数交易
-	MerchantId   int    `json:"merchantId,omitempty"` // 平台商户号,渠道对接可使用merchantId和terminalId参数交易
-	TerminalId   int    `json:"terminalId,omitempty"` // 平台终端号,渠道对接可使用merchantId和terminalId参数交易
+	MerchantId   int64  `json:"merchantId,omitempty"` // 平台商户号,渠道对接可使用merchantId和terminalId参数交易
+	TerminalId   int64  `json:"terminalId,omitempty"` // 平台终端号,渠道对接可使用merchantId和terminalId参数交易
 	SplitOrderId int64  `json:"splitOrderId"`         // 交易分账流水号
 }
